📅 Weather Data

Several (and more coming) weather station format can be read and transformed to DataFrame.

ECA dataset

From the European Climate Assessment & Dataset project at this link for zip of all stations per variables and at this link for custom manual query. I asked them about an API to extract directly a specific file automatically, but they answer it is not currently available. I tried unzip-http but could not get it working with ECA website[1]

using StochasticWeatherGenerators, DataFrames, Dates
collect_data_ECA(33, Date(1956), Date(2019, 12, 31), "https://raw.githubusercontent.com/dmetivie/StochasticWeatherGenerators.jl/master/weather_files/ECA_blend_rr/RR_", portion_valid_data=1, skipto=22, header=21, url=true)[1:10,:]

10×5 DataFrame

Row	STAID	SOUID	DATE	RR	Q_RR

Int64	Int64	Date	Int64	Int64

1	33	105	1956-01-01	23	0

2	33	105	1956-01-02	1	0

3	33	105	1956-01-03	0	0

4	33	105	1956-01-04	0	0

5	33	105	1956-01-05	0	0

6	33	105	1956-01-06	21	0

7	33	105	1956-01-07	3	0

8	33	105	1956-01-08	25	0

9	33	105	1956-01-09	20	0

10	33	105	1956-01-10	0	0

StochasticWeatherGenerators.collect_data_ECA — Function

collect_data_ECA(STAID::Integer, path::String, var::String="RR"; skipto=19, header = 18)

path gives the path where all data files are stored in a vector

collect_data_ECA(STAID, date_start::Date, date_end::Date, path::String, var::String="RR"; portion_valid_data=1, skipto=19, header = 18, return_nothing = true)

path gives the path where all data files are stored in a vector

Filter the DataFrame s.t. date_start ≤ :DATE ≤ date_end

var = "RR", "TX" etc.

portion_valid_data is the portion of valid data we are ok with. If we don't want any missing, fix it to 1.

skipto and header for csv files with meta informations/comments at the beginning of files. See CSV.jl.

return_nothing if true it will return nothing is the file does not exists or does not have enough valid data.

Météo France

Météo France do have a version of this data and it is accessible through an API on the website Data.Gouv.fr. This package provides a simple command to extract the data of one station (given its STAtionID) from the API.

collect_data_MeteoFrance(34154001)[1:10,:] # Montpellier Airport

10×61 DataFrame

Row	__id	STAID	STANAME	LAT	LON	ALTI	DATE	RR	QRR	TN	QTN	HTN	QHTN	TX	QTX	HTX	QHTX	TM	QTM	TNTXM	QTNTXM	TAMPLI	QTAMPLI	TNSOL	QTNSOL	TN50	QTN50	DG	QDG	FFM	QFFM	FF2M	QFF2M	FXY	QFXY	DXY	QDXY	HXY	QHXY	FXI	QFXI	DXI	QDXI	HXI	QHXI	FXI2	QFXI2	DXI2	QDXI2	HXI2	QHXI2	FXI3S	QFXI3S	DXI3S	QDXI3S	HXI3S	QHXI3S	DRR	QDRR	STATUS_FXI3S	STATUS_DXI3S

Int64	Int64	String31	Float64	Float64	Int64	Date	Float64	Int64	Float64	Int64	Int64	Int64	Float64	Int64	Int64	Int64	Float64	Int64	Float64	Int64	Float64	Int64	Float64	Int64	Float64	Int64	Int64?	Int64?	Float64?	Int64	Missing	Missing	Float64?	Int64	Int64?	Int64	Int64?	Int64	Float64?	Int64	Int64?	Int64	Int64?	Int64	Missing	Missing	Missing	Missing	Missing	Missing	Float64?	Int64	Int64?	Int64?	Int64?	Int64	Int64?	Int64?	String1?	String1?

1	4822	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-01	0.0	1	0.7	1	2141	9	8.8	1	1352	9	6.9	1	4.8	1	8.1	1	-3.1	9	-0.9	9	0	9	2.6	1	missing	missing	3.9	1	10	1	131	9	8.2	1	20	1	515	9	missing	missing	missing	missing	missing	missing	6.8	1	20	1	515	9	0	9	f	t

2	4823	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-02	0.0	1	7.0	1	407	9	13.6	1	1349	9	9.8	1	10.3	1	6.6	1	5.8	9	6.2	9	0	9	3.3	1	missing	missing	6.1	1	20	1	908	9	8.2	1	20	1	923	9	missing	missing	missing	missing	missing	missing	7.5	1	20	1	910	9	26	9	f	t

3	4824	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-03	0.0	1	7.4	1	715	9	12.7	1	1322	9	8.6	1	10.1	1	5.3	1	5.6	9	6.4	9	0	9	3.1	1	missing	missing	7.8	1	290	1	149	9	11.0	1	290	1	145	9	missing	missing	missing	missing	missing	missing	10.7	1	290	1	145	9	0	9	f	t

4	4825	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-04	0.0	1	1.5	1	2329	9	7.9	1	1549	9	6.0	1	4.7	1	6.4	1	-1.4	9	-0.1	9	0	9	2.7	1	missing	missing	5.3	1	10	1	317	9	8.5	1	10	1	251	9	missing	missing	missing	missing	missing	missing	7.7	1	10	1	251	9	73	9	f	t

5	4826	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-05	0.0	1	5.8	1	55	9	14.4	1	1302	9	10.7	1	10.1	1	8.6	1	5.8	9	5.8	9	0	9	7.4	1	missing	missing	12.3	1	130	9	2358	9	17.0	1	120	9	2358	9	missing	missing	missing	missing	missing	missing	15.8	1	120	1	2358	9	0	9	f	t

6	4827	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-06	0.2	1	12.5	1	1801	9	16.0	1	1428	9	13.5	1	14.3	1	3.5	1	11.5	9	12.3	9	0	9	6.3	1	missing	missing	12.6	1	130	1	122	9	16.3	1	130	1	9	9	missing	missing	missing	missing	missing	missing	14.9	1	130	1	10	9	36	9	f	t

7	4828	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-07	0.0	1	8.0	1	515	9	14.7	1	1418	9	10.7	1	11.4	1	6.7	1	4.8	9	6.3	9	0	9	4.5	1	missing	missing	10.0	1	300	1	1106	9	14.7	1	300	1	1105	9	missing	missing	missing	missing	missing	missing	14.0	1	300	1	1110	9	49	9	f	t

8	4829	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-08	0.0	1	5.7	1	639	9	14.6	1	1252	9	10.5	1	10.2	1	8.9	1	1.9	9	3.4	9	0	9	3.4	1	missing	missing	5.9	1	80	1	1538	9	9.6	1	50	9	2341	9	missing	missing	missing	missing	missing	missing	8.4	1	80	1	1442	9	0	9	f	t

9	4830	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-09	0.2	1	5.5	1	710	9	16.8	1	1542	9	11.3	1	11.2	1	11.3	1	1.9	9	3.1	9	0	9	5.3	1	missing	missing	14.8	1	300	9	2052	9	23.0	1	310	9	2043	9	missing	missing	missing	missing	missing	missing	20.8	1	300	1	2137	9	0	9	f	t

10	4831	34154001	MONTPELLIER-AEROPORT	43.5762	3.96467	1	2025-01-10	0.0	1	5.1	1	407	9	13.4	1	2103	9	9.3	1	9.3	1	8.3	1	1.9	9	3.6	9	0	9	2.9	1	missing	missing	8.7	1	300	1	44	9	12.3	1	300	1	44	9	missing	missing	missing	missing	missing	missing	11.9	1	300	1	44	9	69	9	f	t

Warning

As it is rather new, this DataGov/MeteoFrance API may change in the future making this function not working anymore. It is currently not fully working. One would have to call the DataGov API directly.

StochasticWeatherGenerators.collect_data_MeteoFrance — Function

collect_data_MeteoFrance(STAID; show_warning=false, impute_missing=[], period="1950-2021", variables = "all")

Given a STAID (station ID given by Météo France), it returns a DataFrame with data in period and for the variables.

STAID can be an integer or string.

Option for period are "1846-1949", "1950-2021", "2022-2023"

Option for variables are all, "RR-T-Wind", "others"

impute_missing expects a vector of column name(s) where to impute missing with Impute.Interpolate e.g. impute_missing=[:TX].

show_warning in case of missing data. false for no column, true for all variables columns and for selected columns e.g. show_warning = [:TX].

The data is available through the French Data.gouv.fr website api. Data may be updated without notice. See the following two links to get informations on the "RR-T-Wind" and "others" variables (in French)

https://object.files.data.gouv.fr/meteofrance/data/synchroftp/BASE/QUOT/QdescriptifchampsRR-T-Vent.csv

https://object.files.data.gouv.fr/meteofrance/data/synchroftp/BASE/QUOT/Qdescriptifchampsautres-parametres.csv

Or the the SICLIMA website with information (in French) about computation and conversion for some weather variables/index.

StochasticWeatherGenerators.download_data_MeteoFrance — Function

download_data_MeteoFrance(STAID, period = "2024-2025", variables = "all")

Function not really working anymore as the API changed in 2024.

Option for period are "1846-1949", "1950-2021", "2022-2023"

Option for variables are all, "RR-T-Wind", "others"

The data is available through the French Data.gouv.fr website api. Data may be updated without notice. In particular the path to the data may change.

INRAE

The INRAE CLIMATIK platform (Delannoy et al., 2022) (https://agroclim.inrae.fr/climatik/, in French) managed by the AgroClim laboratory of Avignon, France has weather stations. However, their API is not open access.

StochasticWeatherGenerators.collect_data_INRAE — Function

collect_data_INRAE(station_path::String; show_warning=false, impute_missing=[])

Read from a file an INRAE formatted weather station data and transform it to match ECA standard naming conventions.

impute_missing expects a vector of column name(s) where to impute missing with Impute.Interpolate e.g. impute_missing=[:TX].

show_warning in case of missing data. false for no column, true for all variables columns and for selected columns e.g. show_warning = [:TX].

Others

Data manipulation

StochasticWeatherGenerators.clean_data — Function

clean_data(df::DataFrame; show_warning=false, impute_missing=[])

Impute missing and show warning for missings. It assumes that the first two columns are not numeric.

impute_missing expects a vector of column name(s) where to impute missing with Impute.Interpolate e.g. impute_missing=[:TX].

show_warning in case of missing data. false for no column, true for all variables columns and for selected columns e.g. show_warning = [:TX].

StochasticWeatherGenerators.select_in_range_df — Function

select_in_range_df(datas, start_Date, interval_Date, [portion])

Select station with some data availability in dates and quality (portion of valid data). Input is a vector (array) of DataFrame (one for each station for example) or a Dict of DataFrame. If 0 < portion ≤ 1 is specified, it will authorize some portion of data to be missing.

StochasticWeatherGenerators.shortname — Function

shortname(name::String)

Experimental function that returns only the most relevant part of a station name.

long_name = "TOULOUSE-BLAGNAC"
shortname(long_name) # "TOULOUSE"

References

Delannoy, D.; Maury, O. and Décome, J. (2022). CLIMATIK: système d’information pour les données du réseau agroclimatique INRAE.

1I don't remember exactly in fact.